Major refactor and move request handlers
