+++ .env.default
@@ -16,3 +16,4 @@ infuraNetwork="rinkeby"
 githubPagesDomain="exokit.dev"
 publicIp="13.57.56.187"
 privateIp="172.31.18.137"
+awsRegion="us-west-2"
