+++ src/constants.js
@@ -12,10 +12,12 @@ const accountKeys = [
   'ftu',
   'addressProofs',
 ];
+
 const ids = {
   lastCachedBlockAccount: 'lastCachedBlock',
   lastCachedBlockNft: -1,
 };
+
 const tableNames = {
   mainnetAccount: 'mainnet-cache-account',
   mainnetNft: 'mainnet-cache-nft',
@@ -30,6 +32,7 @@ const tableNames = {
   testnetpolygonAccount: 'testnetpolygon-cache-account',
   testnetpolygonNft: 'testnetpolygon-cache-nft',
 };
+
 const redisPrefixes = (() => {
   const result = {};
   for (const k in tableNames) {
@@ -37,12 +40,21 @@ const redisPrefixes = (() => {
   }
   return result;
 })();
+
 const nftIndexName = 'nftIdx';
 const polygonVigilKey = `1bdde9289621d9d420488a9804254f4a958e128b`;
 const ethereumHost = 'ethereum.exokit.org';
 const storageHost = 'https://ipfs.exokit.org';
 const mainnetSignatureMessage = `Connecting mainnet address.`;
 const cacheHostUrl = 'cache.webaverse.com';
+
+const userTableName = 'users';
+const defaultAvatarPreview = `https://preview.exokit.org/[https://raw.githubusercontent.com/avaer/vrm-samples/master/vroid/male.vrm]/preview.png`;
+const emailRegex = /(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])/;
+const codeTestRegex = /^[0-9]{6}$/;
+const discordIdTestRegex = /^[0-9]+$/;
+const twitterIdTestRegex = /^@?(\w){1,15}$/;
+
 module.exports = {
   accountKeys,
   ids,
@@ -54,4 +66,10 @@ module.exports = {
   storageHost,
   mainnetSignatureMessage,
   cacheHostUrl,
+  userTableName,
+  defaultAvatarPreview,
+  emailRegex,
+  codeTestRegex,
+  discordIdTestRegex,
+  twitterIdTestRegex
 };
