+++ src/routes/unlock.js
@@ -4,7 +4,7 @@ const fetch = require('node-fetch');
 const Web3 = require('web3');
 const bip39 = require('bip39');
 const { hdkey } = require('ethereumjs-wallet');
-const { jsonParse, _setCorsHeaders } = require('../utils.js');
+const { jsonParse, setCorsHeaders } = require('../utils.js');
 const { polygonVigilKey } = require('../constants.js');
 
 let config = require('fs').existsSync('../../config.json') ? require('../../config.json') : null;
@@ -221,7 +221,7 @@ const _handleUnlockRequest = async (req, res) => {
 
   const request = url.parse(req.url);
   try {
-    res = _setCorsHeaders(res);
+    res = setCorsHeaders(res);
     const { method } = req;
     if (method === 'OPTIONS') {
       res.end();
