+++ src/utils.js
@@ -5,16 +5,22 @@ const jsonParse = (s, d = null) => {
     return d;
   }
 };
-const _setCorsHeaders = res => {
+const setCorsHeaders = res => {
     res.setHeader('Access-Control-Allow-Origin', '*');
     res.setHeader('Access-Control-Allow-Headers', '*');
     res.setHeader('Access-Control-Allow-Methods', '*');
     return res;
 };
+
+function randomString() {
+  return Math.random().toString(36).replace(/[^a-z]+/g, '').substr(0, 5);
+}
+
 function getExt(fileName) {
   const match = fileName.match(/\.([^\.]+)$/);
   return match && match[1].toLowerCase();
 }
+
 const makePromise = () => {
   let accept, reject;
   const p = new Promise((a, r) => {
@@ -27,7 +33,8 @@ const makePromise = () => {
 };
 module.exports = {
   jsonParse,
-  _setCorsHeaders,
+  setCorsHeaders,
   getExt,
   makePromise,
+  randomString
 }
